eye/requirements-ci.txt
# Lightweight CI dependencies (excludes torch, sentence-transformers, mecab)
neo4j>=5.0.0
networkx>=3.0
numpy>=1.24.0
openai>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0
jinja2>=3.1.0
rich>=13.0.0
typer>=0.9.0
orjson>=3.9.0
tiktoken>=0.5.0
python-igraph>=0.11.0
leidenalg>=0.10.0
kss>=6.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
httpx>=0.24.0
pytest-asyncio>=0.21.0
